Firmware/tests/integration/harness/requirements.txt
pytest>=8.0
python-can>=4.5
